backend/requirements.txt
# 📋 Backend Requirements
# Python dependencies for the video streaming platform backend with database support

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
aiofiles==23.2.1

# Database Support
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Video Processing
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
pillow==10.1.0

# Monitoring & Metrics
prometheus-client==0.19.0
psutil==5.9.6
pynvml==11.5.0

# Database & Caching
redis==5.0.1
pymongo==4.6.0
hiredis==2.2.3

# Security & Authentication
cryptography==41.0.7
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Network & HTTP
httpx==0.25.2
requests==2.31.0
websockets==12.0

# Configuration & Logging
pyyaml==6.0.1
python-multipart==0.0.6
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Cloud SDKs
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# CDN Integration
cloudflare==2.19.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
